requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core>=0.3.59
langgraph>=0.4.7
langgraph-checkpoint>=2.0.26
langgraph-prebuilt>=0.2.1
langgraph-sdk>=0.1.70
langsmith==0.3.18
msgpack==1.1.0
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
# Speech recognition components
pyaudio==0.2.14
pywhispercpp==1.3.0
python-dotenv==1.0.1
# VAD components
torch>=2.2.0
torchaudio>=2.2.0
silero-vad>=0.4.0
# WebSocket server
websockets>=12.0.0
# TTS components
kokoro>=0.7.4
soundfile>=0.13.1
langchain-mcp-adapters>=0.1.1
langchain-community>=0.3.24
langchain-openai>=0.3.18
langgraph-cli[inmem]
# Spotify Web API
spotipy>=2.25.1
